requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
arq>=0.25.0
redis>=5.0.0
httpx>=0.27.0
supabase>=2.3.0
pydantic>=2.7.0
python-dotenv>=1.0.0
xgboost>=2.0.0
pandas>=2.2.0
